Introduce a coreboot Control Block (CCB)
It is annoying to have to create and maintain two completely
different builds of coreboot just make minor changes, such as to enable
or disable the console.

It would be much more convenient to have a 'silent' flag in the
image, which can be updated as needed, without needing to rebuild
coreboot.

Introduce the 'coreboot Control Block' (CCB) which can hold such
settings. It is designed to be available very early in bootblock,
before CBFS is ready. It is able to control the output of the very
first bootblock banner.

The CCB is then passed through the cache and placed in cbmem so it is
available to other stages.

Provide options in cbfstool to get and set settings in the CCB. This
makes it easy to use this feature.

BUG=b:172341184, b:262546009, b:249105972
BRANCH=none
TEST=see next CL

Change-Id: I04e946b33035a493e833500351a0483761252613
Signed-off-by: Simon Glass <sjg@chromium.org>